validation tests passing with type check functions moved to class methods
