+++ src/natcap/invest/validation.py
@@ -7,13 +7,8 @@ import inspect
 import logging
 import os
 import pprint
-import queue
-import re
-import threading
-import warnings
 
 import numpy
-import pandas
 import pint
 import pygeoprocessing
 from osgeo import gdal
@@ -21,7 +16,6 @@ from osgeo import ogr
 from osgeo import osr
 
 from . import gettext
-from . import spec_utils
 from . import utils
 
 #: A flag to pass to the validation context manager indicating that all keys
@@ -158,171 +152,6 @@ def get_sufficient_keys(args):
     return sufficient_keys
 
 
-def check_directory(dirpath, must_exist=True, permissions='rx', **kwargs):
-    """Validate a directory.
-
-    Args:
-        dirpath (string): The directory path to validate.
-        must_exist=True (bool): If ``True``, the directory at ``dirpath``
-            must already exist on the filesystem.
-        permissions='rx' (string): A string that includes the lowercase
-            characters ``r``, ``w`` and/or ``x``, indicating read, write, and
-            execute permissions (respectively) required for this directory.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    if must_exist:
-        if not os.path.exists(dirpath):
-            return MESSAGES['DIR_NOT_FOUND']
-
-    if os.path.exists(dirpath):
-        if not os.path.isdir(dirpath):
-            return MESSAGES['NOT_A_DIR']
-    else:
-        # find the parent directory that does exist and check permissions
-        child = dirpath
-        parent = os.path.normcase(os.path.abspath(dirpath))
-        while child:
-            # iterate child because if this gets back to the root dir,
-            # child becomes an empty string and parent remains root string.
-            parent, child = os.path.split(parent)
-            if os.path.exists(parent):
-                dirpath = parent
-                break
-
-    MESSAGE_KEY = 'NEED_PERMISSION_DIRECTORY'
-
-    if 'r' in permissions:
-        try:
-            os.scandir(dirpath).close()
-        except OSError:
-            return MESSAGES[MESSAGE_KEY].format(permission='read')
-
-    # Check for x access before checking for w,
-    # since w operations to a dir are dependent on x access
-    if 'x' in permissions:
-        try:
-            cwd = os.getcwd()
-            os.chdir(dirpath)
-        except OSError:
-            return MESSAGES[MESSAGE_KEY].format(permission='execute')
-        finally:
-            os.chdir(cwd)
-
-    if 'w' in permissions:
-        try:
-            temp_path = os.path.join(dirpath, 'temp__workspace_validation.txt')
-            with open(temp_path, 'w') as temp:
-                temp.close()
-                os.remove(temp_path)
-        except OSError:
-            return MESSAGES[MESSAGE_KEY].format(permission='write')
-
-
-def check_file(filepath, permissions='r', **kwargs):
-    """Validate a single file.
-
-    Args:
-        filepath (string): The filepath to validate.
-        permissions='r' (string): A string that includes the lowercase
-            characters ``r``, ``w`` and/or ``x``, indicating read, write, and
-            execute permissions (respectively) required for this file.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    if not os.path.exists(filepath):
-        return MESSAGES['FILE_NOT_FOUND']
-
-    for letter, mode, descriptor in (
-            ('r', os.R_OK, 'read'),
-            ('w', os.W_OK, 'write'),
-            ('x', os.X_OK, 'execute')):
-        if letter in permissions and not os.access(filepath, mode):
-            return MESSAGES['NEED_PERMISSION_FILE'].format(permission=descriptor)
-
-
-def _check_projection(srs, projected, projection_units):
-    """Validate a GDAL projection.
-
-    Args:
-        srs (osr.SpatialReference): A GDAL Spatial Reference object
-            representing the spatial reference of a GDAL dataset.
-        projected (bool): Whether the spatial reference must be projected in
-            linear units.
-        projection_units (pint.Unit): The projection's required linear units.
-
-    Returns:
-        A string error message if an error was found. ``None`` otherwise.
-
-    """
-    empty_srs = osr.SpatialReference()
-    if srs is None or srs.IsSame(empty_srs):
-        return MESSAGES['INVALID_PROJECTION']
-
-    if projected:
-        if not srs.IsProjected():
-            return MESSAGES['NOT_PROJECTED']
-
-    if projection_units:
-        # pint uses underscores in multi-word units e.g. 'survey_foot'
-        # it is case-sensitive
-        layer_units_name = srs.GetLinearUnitsName().lower().replace(' ', '_')
-        try:
-            # this will parse common synonyms: m, meter, meters, metre, metres
-            layer_units = spec_utils.u.Unit(layer_units_name)
-            # Compare pint Unit objects
-            if projection_units != layer_units:
-                return MESSAGES['WRONG_PROJECTION_UNIT'].format(
-                    unit_a=projection_units, unit_b=layer_units_name)
-        except pint.errors.UndefinedUnitError:
-            return MESSAGES['WRONG_PROJECTION_UNIT'].format(
-                unit_a=projection_units, unit_b=layer_units_name)
-
-    return None
-
-
-def check_raster(filepath, projected=False, projection_units=None, **kwargs):
-    """Validate a GDAL Raster on disk.
-
-    Args:
-        filepath (string): The path to the raster on disk.  The file must exist
-            and be readable.
-        projected=False (bool): Whether the spatial reference must be projected
-            in linear units.
-        projection_units=None (pint.Units): The required linear units of the
-            projection. If ``None``, the projection units will not be checked.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    file_warning = check_file(filepath, permissions='r')
-    if file_warning:
-        return file_warning
-
-    try:
-        gdal_dataset = gdal.OpenEx(filepath, gdal.OF_RASTER)
-    except RuntimeError:
-        return MESSAGES['NOT_GDAL_RASTER']
-
-    # Check that an overview .ovr file wasn't opened.
-    if os.path.splitext(filepath)[1] == '.ovr':
-        return MESSAGES['OVR_FILE']
-
-    srs = gdal_dataset.GetSpatialRef()
-    projection_warning = _check_projection(srs, projected, projection_units)
-    if projection_warning:
-        gdal_dataset = None
-        return projection_warning
-
-    gdal_dataset = None
-    return None
-
-
 def load_fields_from_vector(filepath, layer_id=0):
     """Load fieldnames from a given vector.
 
@@ -343,431 +172,6 @@ def load_fields_from_vector(filepath, layer_id=0):
     layer = None
     vector = None
     return fieldnames
-
-
-def check_vector(filepath, geometries, fields=None, projected=False,
-                 projection_units=None, **kwargs):
-    """Validate a GDAL vector on disk.
-
-    Note:
-        If the provided vector has multiple layers, only the first layer will
-        be checked.
-
-    Args:
-        filepath (string): The path to the vector on disk.  The file must exist
-            and be readable.
-        geometries (set): Set of geometry type(s) that are allowed. Options are
-            'POINT', 'LINESTRING', 'POLYGON', 'MULTIPOINT', 'MULTILINESTRING',
-            and 'MULTIPOLYGON'.
-        fields=None (dict): A dictionary spec of field names that the vector is
-            expected to have. See the docstring of ``check_headers`` for
-            details on validation rules.
-        projected=False (bool): Whether the spatial reference must be projected
-            in linear units.  If None, the projection will not be checked.
-        projection_units=None (pint.Units): The required linear units of the
-            projection. If ``None``, the projection units will not be checked.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    file_warning = check_file(filepath, permissions='r')
-    if file_warning:
-        return file_warning
-
-    try:
-        gdal_dataset = gdal.OpenEx(filepath, gdal.OF_VECTOR)
-    except RuntimeError:
-        return MESSAGES['NOT_GDAL_VECTOR']
-
-    geom_map = {
-        'POINT': [ogr.wkbPoint, ogr.wkbPointM, ogr.wkbPointZM,
-                  ogr.wkbPoint25D],
-        'LINESTRING': [ogr.wkbLineString, ogr.wkbLineStringM,
-                       ogr.wkbLineStringZM, ogr.wkbLineString25D],
-        'POLYGON': [ogr.wkbPolygon, ogr.wkbPolygonM,
-                    ogr.wkbPolygonZM, ogr.wkbPolygon25D],
-        'MULTIPOINT': [ogr.wkbMultiPoint, ogr.wkbMultiPointM,
-                       ogr.wkbMultiPointZM, ogr.wkbMultiPoint25D],
-        'MULTILINESTRING': [ogr.wkbMultiLineString, ogr.wkbMultiLineStringM,
-                            ogr.wkbMultiLineStringZM,
-                            ogr.wkbMultiLineString25D],
-        'MULTIPOLYGON': [ogr.wkbMultiPolygon, ogr.wkbMultiPolygonM,
-                         ogr.wkbMultiPolygonZM, ogr.wkbMultiPolygon25D]
-    }
-
-    allowed_geom_types = []
-    for geom in geometries:
-        allowed_geom_types += geom_map[geom]
-
-    # NOTE: this only checks the layer geometry type, not the types of the
-    # actual geometries (layer.GetGeometryTypes()). This is probably equivalent
-    # in most cases, and it's more efficient than checking every geometry, but
-    # we might need to change this in the future if it becomes a problem.
-    # Currently not supporting ogr.wkbUnknown which allows mixed types.
-    layer = gdal_dataset.GetLayer()
-    if layer.GetGeomType() not in allowed_geom_types:
-        return MESSAGES['WRONG_GEOM_TYPE'].format(allowed=geometries)
-
-    if fields:
-        field_patterns = get_headers_to_validate(fields)
-        fieldnames = [defn.GetName() for defn in layer.schema]
-        required_field_warning = check_headers(
-            field_patterns, fieldnames, 'field')
-        if required_field_warning:
-            return required_field_warning
-
-    srs = layer.GetSpatialRef()
-    projection_warning = _check_projection(srs, projected, projection_units)
-    return projection_warning
-
-
-def check_raster_or_vector(filepath, **kwargs):
-    """Validate an input that may be a raster or vector.
-
-    Args:
-        filepath (string):  The path to the raster or vector.
-        **kwargs: kwargs of the raster and vector spec. Will be
-            passed to ``check_raster`` or ``check_vector``.
-
-    Returns:
-        A string error message if an error was found. ``None`` otherwise.
-    """
-    try:
-        gis_type = pygeoprocessing.get_gis_type(filepath)
-    except ValueError as err:
-        return str(err)
-    if gis_type == pygeoprocessing.RASTER_TYPE:
-        return check_raster(filepath, **kwargs)
-    else:
-        return check_vector(filepath, **kwargs)
-
-
-def check_freestyle_string(value, regexp=None, **kwargs):
-    """Validate an arbitrary string.
-
-    Args:
-        value: The value to check.  Must be able to be cast to a string.
-        regexp=None (string): a string interpreted as a regular expression.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    if regexp:
-        matches = re.fullmatch(regexp, str(value))
-        if not matches:
-            return MESSAGES['REGEXP_MISMATCH'].format(regexp=regexp)
-    return None
-
-
-def check_option_string(value, options, **kwargs):
-    """Validate that a string is in a set of options.
-
-    Args:
-        value: The value to test. Will be cast to a string before comparing
-            against the allowed options.
-        options (dict): option spec to validate against.
-
-    Returns:
-        A string error message if ``value`` is not in ``options``.  ``None``
-        otherwise.
-
-    """
-    # if options is empty, that means it's dynamically populated
-    # so validation should be left to the model's validate function.
-    if options and str(value) not in options:
-        return MESSAGES['INVALID_OPTION'].format(option_list=sorted(options))
-
-
-def check_number(value, expression=None, **kwargs):
-    """Validate numbers.
-
-    Args:
-        value: A python value. This should be able to be cast to a float.
-        expression=None (string): A string expression to be evaluated with the
-            intent of determining that the value is within a specific range.
-            The expression must contain the string ``value``, which will
-            represent the user-provided value (after it has been cast to a
-            float).  Example expression: ``"(value >= 0) & (value <= 1)"``.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    try:
-        float(value)
-    except (TypeError, ValueError):
-        return MESSAGES['NOT_A_NUMBER'].format(value=value)
-
-    if expression:
-        # Check to make sure that 'value' is in the expression.
-        if 'value' not in expression:
-            raise AssertionError(
-                'The variable name value is not found in the '
-                f'expression: {expression}')
-
-        # Expression is assumed to return a boolean, something like
-        # "value > 0" or "(value >= 0) & (value < 1)".  An exception will
-        # be raised if asteval can't evaluate the expression.
-        result = _evaluate_expression(expression, {'value': float(value)})
-        if not result:  # A python bool object is returned.
-            return MESSAGES['INVALID_VALUE'].format(condition=expression)
-
-    return None
-
-
-def check_ratio(value, **kwargs):
-    """Validate a ratio (a proportion expressed as a value from 0 to 1).
-
-    Args:
-        value: A python value. This should be able to be cast to a float.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    try:
-        as_float = float(value)
-    except (TypeError, ValueError):
-        return MESSAGES['NOT_A_NUMBER'].format(value=value)
-
-    if as_float < 0 or as_float > 1:
-        return MESSAGES['NOT_WITHIN_RANGE'].format(
-            value=as_float,
-            range='[0, 1]')
-
-    return None
-
-
-def check_percent(value, **kwargs):
-    """Validate a percent (a proportion expressed as a value from 0 to 100).
-
-    Args:
-        value: A python value. This should be able to be cast to a float.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    try:
-        as_float = float(value)
-    except (TypeError, ValueError):
-        return MESSAGES['NOT_A_NUMBER'].format(value=value)
-
-    if as_float < 0 or as_float > 100:
-        return MESSAGES['NOT_WITHIN_RANGE'].format(
-            value=as_float,
-            range='[0, 100]')
-
-    return None
-
-
-def check_integer(value, **kwargs):
-    """Validate an integer.
-
-    Args:
-        value: A python value. This should be able to be cast to an int.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    try:
-        # must first cast to float, to handle both string and float inputs
-        as_float = float(value)
-        if not as_float.is_integer():
-            return MESSAGES['NOT_AN_INTEGER'].format(value=value)
-    except (TypeError, ValueError):
-        return MESSAGES['NOT_A_NUMBER'].format(value=value)
-    return None
-
-
-def check_boolean(value, **kwargs):
-    """Validate a boolean value.
-
-    If the value provided is not a python boolean, an error message is
-    returned.
-
-
-    Args:
-        value: The value to evaluate.
-
-    Returns:
-        A string error message if an error was found.  ``None`` otherwise.
-
-    """
-    if not isinstance(value, bool):
-        return MESSAGES['NOT_BOOLEAN'].format(value=value)
-
-
-def get_validated_dataframe(
-        csv_path, columns=None, rows=None, index_col=None,
-        read_csv_kwargs={}, **kwargs):
-    """Read a CSV into a dataframe that is guaranteed to match the spec."""
-
-    if not (columns or rows):
-        raise ValueError('One of columns or rows must be provided')
-
-    # build up a list of regex patterns to match columns against columns from
-    # the table that match a pattern in this list (after stripping whitespace
-    # and lowercasing) will be included in the dataframe
-    axis = 'column' if columns else 'row'
-
-    if rows:
-        read_csv_kwargs = read_csv_kwargs.copy()
-        read_csv_kwargs['header'] = None
-
-    df = utils.read_csv_to_dataframe(csv_path, **read_csv_kwargs)
-
-    if rows:
-        # swap rows and column
-        df = df.set_index(df.columns[0]).rename_axis(
-            None, axis=0).T.reset_index(drop=True)
-
-    columns = columns if columns else rows
-
-    patterns = []
-    for column in columns:
-        column = column.id.lower()
-        match = re.match(r'(.*)\[(.+)\](.*)', column)
-        if match:
-            # for column name patterns, convert it to a regex pattern
-            groups = match.groups()
-            patterns.append(f'{groups[0]}(.+){groups[2]}')
-        else:
-            # for regular column names, use the exact name as the pattern
-            patterns.append(column.replace('(', r'\(').replace(')', r'\)'))
-
-    # select only the columns that match a pattern
-    df = df[[col for col in df.columns if any(
-        re.fullmatch(pattern, col) for pattern in patterns)]]
-
-    # drop any empty rows
-    df = df.dropna(how="all").reset_index(drop=True)
-
-    available_cols = set(df.columns)
-
-    for col_spec, pattern in zip(columns, patterns):
-        matching_cols = [c for c in available_cols if re.fullmatch(pattern, c)]
-        if col_spec.required and '[' not in col_spec.id and not matching_cols:
-            raise ValueError(MESSAGES['MATCHED_NO_HEADERS'].format(
-                header=axis,
-                header_name=col_spec.id))
-        available_cols -= set(matching_cols)
-        for col in matching_cols:
-            try:
-                # frozenset needed to make the set hashable.  A frozenset and set with the same members are equal.
-                if type(col_spec) in {spec_utils.CSVInputSpec, spec_utils.DirectoryInputSpec,
-                                      spec_utils.FileInputSpec, spec_utils.SingleBandRasterInputSpec,
-                                      spec_utils.VectorInputSpec, spec_utils.RasterOrVectorInputSpec}:
-                    df[col] = df[col].apply(
-                        lambda p: p if pandas.isna(p) else utils.expand_path(str(p).strip(), csv_path))
-                    df[col] = df[col].astype(pandas.StringDtype())
-                elif type(col_spec) in {spec_utils.StringInputSpec, spec_utils.OptionStringInputSpec}:
-                    df[col] = df[col].apply(
-                        lambda s: s if pandas.isna(s) else str(s).strip().lower())
-                    df[col] = df[col].astype(pandas.StringDtype())
-                elif type(col_spec) in {spec_utils.NumberInputSpec, spec_utils.PercentInputSpec, spec_utils.RatioInputSpec}:
-                    df[col] = df[col].astype(float)
-                elif type(col_spec) == spec_utils.IntegerInputSpec:
-                    df[col] = df[col].astype(pandas.Int64Dtype())
-                elif type(col_spec) == spec_utils.BooleanInputSpec:
-                    df[col] = df[col].astype('boolean')
-                else:
-                    raise ValueError(f"Unknown type: {type(col_spec)}")
-            except Exception as err:
-                raise ValueError(
-                    f'Value(s) in the "{col}" column could not be interpreted '
-                    f'as {type(col_spec)}s. Original error: {err}')
-
-            if type(col_spec) in {spec_utils.SingleBandRasterInputSpec,
-                                  spec_utils.VectorInputSpec, spec_utils.RasterOrVectorInputSpec}:
-                # recursively validate the files within the column
-                def check_value(value):
-                    if pandas.isna(value):
-                        return
-                    err_msg = _VALIDATION_FUNCS[type(col_spec)](value, **col_spec.__dict__)
-                    if err_msg:
-                        raise ValueError(
-                            f'Error in {axis} "{col}", value "{value}": {err_msg}')
-                df[col].apply(check_value)
-
-    if any(df.columns.duplicated()):
-        duplicated_columns = df.columns[df.columns.duplicated]
-        return MESSAGES['DUPLICATE_HEADER'].format(
-            header=header_type,
-            header_name=expected,
-            number=count)
-
-    # set the index column, if specified
-    if index_col is not None:
-        index_col = index_col.lower()
-        try:
-            df = df.set_index(index_col, verify_integrity=True)
-        except KeyError:
-            # If 'index_col' is not a column then KeyError is raised for using
-            # it as the index column
-            LOGGER.error(f"The column '{index_col}' could not be found "
-                         f"in the table {csv_path}")
-            raise
-
-    return df
-
-
-def check_csv(filepath, columns=None, rows=None, **kwargs):
-    """Validate a table.
-
-    Args:
-        filepath (string): The string filepath to the table.
-
-    Returns:
-        A string error message if an error was found. ``None`` otherwise.
-
-    """
-    file_warning = check_file(filepath, permissions='r')
-    if file_warning:
-        return file_warning
-    if columns or rows:
-        try:
-            get_validated_dataframe(filepath, columns=columns, rows=rows, **kwargs)
-        except Exception as e:
-            return str(e)
-
-
-def check_headers(expected_headers, actual_headers, header_type='header'):
-    """Validate that expected headers are in a list of actual headers.
-
-    - Each expected header should be found exactly once.
-    - Actual headers may contain extra headers that are not expected.
-    - Headers are converted to lowercase before matching.
-
-    Args:
-        expected_headers (list[str]): A list of headers that are expected to
-            exist in `actual_headers`.
-        actual_headers (list[str]): A list of actual headers to validate
-            against `expected_headers`.
-        header_type (str): A string to use in the error message to refer to the
-            header (typically one of 'column', 'row', 'field')
-
-    Returns:
-        None, if validation passes; or a string describing the problem, if a
-        validation rule is broken.
-    """
-    actual_headers = [header.lower()
-                      for header in actual_headers]  # case insensitive
-    for expected in expected_headers:
-        count = actual_headers.count(expected)
-        if count == 0:
-            return MESSAGES['MATCHED_NO_HEADERS'].format(
-                header=header_type,
-                header_name=expected)
-        elif count > 1:
-            return MESSAGES['DUPLICATE_HEADER'].format(
-                header=header_type,
-                header_name=expected,
-                number=count)
-    return None
 
 
 def check_spatial_overlap(spatial_filepaths_list,
@@ -832,95 +236,6 @@ def _format_bbox_list(file_list, bbox_list):
     return ' | '.join(
             [a + ': ' + str(b) for a, b in zip(
                 file_list, bbox_list)])
-
-
-def timeout(func, *args, timeout=5, **kwargs):
-    """Stop a function after a given amount of time.
-
-    Args:
-        func (function): function to apply the timeout to
-        args: arguments to pass to the function
-        timeout (number): how many seconds to allow the function to run.
-            Defaults to 5.
-
-    Returns:
-        A string warning message if the thread completed in time and returned
-        warnings, ``None`` otherwise.
-
-    Raises:
-        ``RuntimeWarning`` if the thread does not complete in time.
-    """
-    # use a queue to share the return value from the file checking thread
-    # the target function puts the return value from `func` into shared memory
-    message_queue = queue.Queue()
-
-    def wrapper_func():
-        message_queue.put(func(*args, **kwargs))
-
-    thread = threading.Thread(target=wrapper_func)
-    LOGGER.debug(f'Starting file checking thread with timeout={timeout}')
-    thread.start()
-    thread.join(timeout=timeout)
-    if thread.is_alive():
-        # first arg to `check_csv`, `check_raster`, `check_vector` is the path
-        warnings.warn(
-            f'Validation of file {args[0]} timed out. If this file '
-            'is stored in a file streaming service, it may be taking a long '
-            'time to download. Try storing it locally instead.')
-        return None
-
-    else:
-        LOGGER.debug('File checking thread completed.')
-        # get any warning messages returned from the thread
-        a = message_queue.get()
-        return a
-
-
-def get_headers_to_validate(specs):
-    """Get header names to validate from a row/column/field spec dictionary.
-
-    This module only validates row/column/field names that are static and
-    always required. If `'required'` is anything besides `True`, or if the name
-    contains brackets indicating it's user-defined, it is not returned.
-
-    Args:
-        specs (dict): a row/column/field spec dictionary that maps row/column/
-            field names to specs for them
-
-    Returns:
-        list of expected header names to validate against
-    """
-    headers = []
-    for spec in specs:
-        # if 'required' isn't a key, it defaults to True
-        if spec.required is True:
-            # brackets are a special character for our args spec syntax
-            # they surround the part of the key that's user-defined
-            # user-defined rows/columns/fields are not validated here, so skip
-            if '[' not in spec.id:
-                headers.append(spec.id)
-    return headers
-
-
-# accessing a file could take a long time if it's in a file streaming service
-# to prevent the UI from hanging due to slow validation,
-# set a timeout for these functions.
-_VALIDATION_FUNCS = {
-    spec_utils.BooleanInputSpec: check_boolean,
-    spec_utils.CSVInputSpec: functools.partial(timeout, check_csv),
-    spec_utils.FileInputSpec: functools.partial(timeout, check_file),
-    spec_utils.DirectoryInputSpec: functools.partial(timeout, check_directory),
-    spec_utils.StringInputSpec: check_freestyle_string,
-    spec_utils.NumberInputSpec: check_number,
-    spec_utils.RatioInputSpec: check_ratio,
-    spec_utils.PercentInputSpec: check_percent,
-    spec_utils.IntegerInputSpec: check_integer,
-    spec_utils.OptionStringInputSpec: check_option_string,
-    spec_utils.SingleBandRasterInputSpec: functools.partial(timeout, check_raster),
-    spec_utils.VectorInputSpec: functools.partial(timeout, check_vector),
-    spec_utils.RasterOrVectorInputSpec: functools.partial(timeout, check_raster_or_vector),
-    spec_utils.OtherInputSpec: None
-}
 
 
 def validate(args, spec, spatial_overlap_opts=None):
@@ -1001,13 +316,7 @@ def validate(args, spec, spatial_overlap_opts=None):
             continue
 
         # rewrite parameter_spec for any nested, conditional validity
-        axis_keys = None
-        if type(parameter_spec) == spec_utils.CSVInputSpec:
-            axis_keys = ['columns', 'rows']
-        elif type(parameter_spec) in {spec_utils.VectorInputSpec, spec_utils.RasterOrVectorInputSpec}:
-            axis_keys = ['fields']
-        elif type(parameter_spec) == 'directory':
-            axis_keys = ['contents']
+        axis_keys = set(dir(parameter_spec)).intersection({'columns', 'rows', 'fields', 'contents'})
 
         if axis_keys:
             for axis_key in axis_keys:
@@ -1018,23 +327,15 @@ def validate(args, spec, spatial_overlap_opts=None):
                         nested_spec.required = (
                             bool(_evaluate_expression(
                                 nested_spec.required, expression_values)))
-
-        type_validation_func = _VALIDATION_FUNCS[type(parameter_spec)]
-
-        if type_validation_func is None:
-            # Validation for 'other' type must be performed by the user.
-            continue
         try:
             # pass the entire arg spec into the validation function as kwargs
             # each type validation function allows extra kwargs with **kwargs
-            warning_msg = type_validation_func(args[key], **parameter_spec.__dict__)
+            warning_msg = parameter_spec.validate(args[key])
             if warning_msg:
                 validation_warnings.append(([key], warning_msg))
                 invalid_keys.add(key)
         except Exception:
-            LOGGER.exception(
-                'Error when validating key %s with value %s',
-                key, args[key])
+            LOGGER.exception(f'Error when validating key {key} with value {args[key]}')
             validation_warnings.append(([key], MESSAGES['UNEXPECTED_ERROR']))
 
     # Phase 3: Check spatial overlap if applicable
